// src/app/enums-interfaces/social-network.enum.ts
export enum SocialMediaPlatform {
  YOUTUBE = 'youtube',
  FACEBOOK = 'facebook',
  TIKTOK = 'tiktok',
  INSTAGRAM = 'instagram',
  WHATSAPP = 'whatsapp',
}

export enum SocialMediaType {
  VIDEO = 'video',
  STORY = 'story',
  MESSAGE = 'message',
}

export enum PlatformType {
  FREE = 'free',
  PREMIUM = 'premium',
}